Split server.cpp into 4 files for code clarity
