+++ src/makeSelectAndRunCommand.cpp
@@ -25,7 +25,6 @@ int		Server::commandSelector(Msg msg, Client &client)
         send(client.getSocket(), response.c_str(), response.size(), 0);
 	}
 	else if  (msg.command == "PRIVMSG")
-	
 	{
 		privmsgCommand(msg, client);
 	}
@@ -72,7 +71,7 @@ int		Server::commandSelector(Msg msg, Client &client)
 
 /*
 	1. Parses Irssi protocol strings to Msg Object
-	2. Runs command through commandSelector
+	2. Runs command (in Msg) through commandSelector
 	(Can handle multiple commands sent at once.)
 */
 int    Server::makeSelectAndRunCommand(std::string messages, Client &client)
